+++ src/main/java/teymi15/kassistant/model/Ingredient.java
@@ -31,7 +31,7 @@ public class Ingredient {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private int id;
 
-     private double price; //Using different currency
+     private int price; //Using different currency
 
     @NotNull
     @Size(min = 1, message = "The name of the ingredient must be at least {min} letter.")
@@ -83,7 +83,7 @@ public class Ingredient {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(int price) {
         this.price = price;
     }
